requirements.txt
fastapi
"uvicorn[standard]"
celery
redis
python-dotenv
supabase-py
requests
pyyaml
openai
httpx
pypdf
